requirements.txt
fastapi
uvicorn
faiss-cpu
transformers
torch
pydantic
python-dotenv
